Better wallpaper changer

+++ .scripts/change-wallpaper.sh
@@ -4,11 +4,13 @@ dest=$HOME/pics/wallpaper
 
 cd $HOME/pics/wall
 
-test [ -f $dest ] | rm $dest
-
 file=$(ls | shuf | sxiv -oiqt | head -n 1)
 
-ln $file $dest
+if [ ! -z $file ]; then
+	test [ -f $dest ] | rm $dest
+	ln $file $dest
+fi
+
 feh --bg-fill $dest
